working on full GCODE rewriting

+++ convert/cv_main.py
@@ -3,8 +3,7 @@ from datetime import datetime
 import os
 import argparse
 
-
-#test_file_input = 'c:\\Users\\achan\Desktop\\test_zotrax.cws'
+verbose = True
 
 # Settings table : match between Z-Suite gcode header and expected silce.conf
 master_table = [("Pix per mm X", "xppm", float),
@@ -21,9 +20,98 @@ master_table = [("Pix per mm X", "xppm", float),
                 ("Z Lift Feed Rate", "lift_up_speed", int),
                 ("Z Lift Retract Rate", "lift_down_speed", int)]
 
-# lift_when_finished seems to be 80 all the time
+extra_table = [("Anti Aliasing", "aa", str),
+               ("Anti Aliasing Value", "aa_v", float),
+               ("Flip X", "flip_x", str),
+               ("Flip Y", "flip_y", str),
+               ("Z Lift Feed Rate", "lift_feed", float),
+               ("Z Lift Retract Rate", "lift_retract", float),
+               ("Blanking Layer Time", "wait", int)]
 
-verbose = False
+
+# Header generation for GCODE
+def gen_header(conf, extra, timestr):
+    lines = [
+        "# nova3d.cn NovaMaker v2.4.20 64-bits " + timestr,
+        ";(****Build and Slicing Parameters****)",
+        f";(Pix per mm X            = {conf['xppm']} )",
+        f";(Pix per mm Y            = {conf['yppm']} )",
+        f";(X Resolution            = {conf['xres']} )",
+        f";(Y Resolution            = {conf['yres']} )",
+        f";(Layer Thickness         = {conf['thickness']} mm )",
+        f";(Layer Time              = {conf['layers_expo_ms']} ms )",
+        ";(Render Outlines         = False",
+        ";(Outline Width Inset     = 2",
+        ";(Outline Width Outset    = 0",
+        f";(Bottom Layers Time      = {conf['head_layers_expo_ms']} ms )",
+        f";(Number of Bottom Layers = {conf['head_layers_num']} )",
+        ";(Blanking Layer Time     = 2000 ms )",
+        ";(Build Direction         = Bottom_Up)",
+        f";(Lift Distance           = {conf['lift_distance']} mm )",
+        ";(Slide/Tilt Value        = 0)",
+        ";(Use Mainlift GCode Tab  = False)",
+        f";(Anti Aliasing           = {extra['aa']})",
+        f";(Anti Aliasing Value     = {extra['aa_v']} )",
+        f";(Z Lift Feed Rate        = {extra['lift_feed']} mm/s )",
+        f";(Z Bottom Lift Feed Rate = {40.0} mm/s )",
+        f";(Z Lift Retract Rate     = {extra['lift_retract']} mm/s )",
+        f";(Flip X                  = {extra['flip_x']})",
+        f";(Flip Y                  = {extra['flip_y']})",
+        f";Number of Slices         = {conf['layers_num']}",
+        ";(****Machine Configuration ******)",
+        ";(Platform X Size         = 65.02mm )",
+        ";(Platform Y Size         = 116mm )",
+        ";(Platform Z Size         = 130mm )",
+        ";(Max X Feedrate          = 200mm/s )",
+        ";(Max Y Feedrate          = 200mm/s )",
+        ";(Max Z Feedrate          = 200mm/s )",
+        ";(Machine Type            = UV_LCD)",
+        "",
+        "G28",
+        "G21 ;Set units to be mm",
+        "G91 ;Relative Positioning",
+        "M17 ;Enable motors",
+        "<Slice> Blank",
+        "M106 S0",
+        ""
+    ]
+
+    return lines
+
+
+def gen_slices(conf, extra):
+    lines = []
+    slice_idx = 0
+    lift = float(conf['lift_distance'])
+    down = lift - conf['thickness']
+
+    while slice_idx < conf["head_layers_num"]:
+        lines.append(f";<Slice> {slice_idx}")
+        lines.append("M106 S255")
+        lines.append(f";<Delay> {conf['head_layers_expo_ms']}")
+        lines.append("M106 S0")
+        lines.append(";<Slice> Blank")
+        lines.append(f"G1 Z{lift} F20")
+        lines.append(f"G1 Z-{down} F20")
+        lines.append(";<Delay> 10000")
+        lines.append("")
+
+        slice_idx += 1
+
+    while slice_idx < conf['layers_num']:
+        lines.append(f";<Slice> {slice_idx}")
+        lines.append("M106 S255")
+        lines.append(f";<Delay> {conf['layers_expo_ms']}")
+        lines.append("M106 S0\n;<Slice> Blank")
+        lines.append(f"G1 Z{lift} F{extra['lift_feed']}")
+        lines.append(f"G1 Z-{down} F{extra['lift_retract']}")
+        lines.append(f";<Delay> {extra['wait']}")
+        lines.append("")
+
+        slice_idx += 1
+
+    return lines
+
 
 def find_raw_line(search_str_, raw_settings_lines_):
     raw_str_ = None
@@ -42,18 +130,22 @@ if __name__ == '__main__':
 
     parser = argparse.ArgumentParser(description='Process a Z-Suite CWS archive to make it compatible with the Nova3D Elfin Printer')
     parser.add_argument('--file', required=True, help='The .cws print file to modify')
-    args = parser.parse_args()
+    #args = parser.parse_args()
+    #test_file_input = args.file
 
-    test_file_input = args.file
-
+    test_file_input = 'd:\\Hero_Forge_Explorer_Demo.cws'
+    
+    canonical_name = None
+    
     # Open original archive
     with ZipFile(test_file_input, mode="r") as zip_input:
-
+        
         zinfo_gcode = None
         # Fetch gcode
         for z_file_info in zip_input.filelist:
             if z_file_info.filename.endswith(".gcode"):
                 zinfo_gcode = z_file_info
+                canonical_name = z_file_info.filename.split(".gcode")[0]
 
         # Get settings (before Header)
         raw_settings_lines = []
@@ -91,12 +183,59 @@ if __name__ == '__main__':
         outlines.append(clean_str)
         if verbose: print(clean_str)
 
-    outlines.append("lift_when_finished      = 80")
+    # Fetch extra parameters (only for Gcode)
+    extra_data = {}
+    for search_str, key, type in extra_table:
+        raw_str = find_raw_line(search_str, raw_settings_lines)
+        clean_str = raw_str.split(' = ')[1].strip(" ms/px")
+        extra_data[key] = type(clean_str)
 
+
+    outlines.append("lift_when_finished      = 80")
+    
+    # Generating Gcode
+    with ZipFile(test_file_input, mode="r") as zip_input:
+
+        zinfo_gcode = None
+        # Fetch gcode
+        for z_file_info in zip_input.filelist:
+            if z_file_info.filename.endswith(".gcode"):
+                zinfo_gcode = z_file_info
+
+        with open(canonical_name + "_tmp.gcode", mode="w", newline="\n") as gcode_out:
+            # Write Header
+            gcode_out.write("\n".join(gen_header(config_data, extra_data, now_str)) + "\n")
+            # Write Slices
+            gcode_out.write("\n".join(gen_slices(config_data, extra_data)))
+            # Write end
+            gcode_out.write("M18 ;Disable Motors\n")
+            gcode_out.write("M106 S0\n")
+            lift_height = (140 - (config_data['thickness'] * config_data['layers_num'])) - 5
+            if lift_height < 0:
+                raise ValueError
+            gcode_out.write(f"G1 Z{lift_height}\n")
+            gcode_out.write(";<Completed>\n")
+
+
+    # Building new archive
     with open("slice_tmp.conf", mode="w") as out:
         out.write('\n'.join(outlines) + '\n')
 
-    with ZipFile(test_file_input, mode="a") as zip:
-        zip.write("slice_tmp.conf", arcname="slice.conf")
+    out_zip = test_file_input + "_clean.cws"
 
-    os.remove("slice_tmp.conf")
+    with ZipFile(out_zip, mode="w") as zip:
+        # Put conf file in
+        zip.write("slice_tmp.conf", arcname="slice.conf")
+        
+        # Put GCODE in
+        zip.write(canonical_name + "_tmp.gcode", arcname=canonical_name + ".gcode")
+        
+        #Copy images
+        with ZipFile(test_file_input, 'r') as zin:
+            for item in zin.infolist():
+                buffer = zin.read(item.filename)
+                if item.filename.endswith("png"):
+                    zip.writestr(item, buffer)
+
+    if not verbose: os.remove("slice_tmp.conf")
+    if not verbose: os.remove(canonical_name + "_tmp.gcode")
